+++ RELEASE-NOTES.md
@@ -42,6 +42,7 @@ END TEMPLATE-->
 * Added `LocalizedEntityCommands`, which are console commands that have the ability to take entity system dependencies.
 * Added `BeginRegistrationRegion` to `IConsoleHost` to allow efficient bulk-registration of console commands.
 * Added `IConsoleHost.RegisterCommand` overload that takes an `IConsoleCommand`.
+* Added a `Finished` boolean to `AnimationCompletedEvent` which allows distinguishing if an animation was removed prematurely or completed naturally.
 
 ### Bugfixes